[~] Moved exam_details_bloc.dart to designated location [~] Formatting
