+++ src/lib/components/exams/new_exam_card.dart
@@ -2,8 +2,8 @@ import 'package:dotted_border/dotted_border.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:provider/src/provider.dart';
-import 'package:schoolexam_correction_ui/blocs/exams/exam_details_bloc.dart';
-import 'package:schoolexam_correction_ui/blocs/exams/exam_details_event.dart';
+import 'package:schoolexam_correction_ui/blocs/exam_details/exam_details.dart';
+import 'package:schoolexam_correction_ui/blocs/exam_details/exam_details_bloc.dart';
 import 'package:schoolexam_correction_ui/components/exams/new_exam_dialog.dart';
 
 class NewExamCard extends StatelessWidget {
