+++ src/lib/main.dart
@@ -12,7 +12,7 @@ import 'package:schoolexam_correction_ui/navigation/school_exam_router_delegate.
 import 'package:schoolexam_correction_ui/repositories/correction_overlay/correction_overlay.dart';
 import 'package:schoolexam_correction_ui/repositories/correction_overlay/database_correction_overlay_repository.dart';
 
-import 'blocs/exams/exam_details_bloc.dart';
+import 'blocs/exam_details/exam_details_bloc.dart';
 import 'blocs/overlay/correction_overlay.dart';
 import 'blocs/remark/remark.dart';
